# packages/contracts/internal/core-api.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/budgets": {
      "get": {
        "tags": [
          "budget-controller"
        ],
        "operationId": "list_11",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BudgetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "budget-controller"
        ],
        "operationId": "create_7",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/budgets/{id}": {
      "delete": {
        "tags": [
          "budget-controller"
        ],
        "operationId": "delete_6",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BudgetRequest": {
        "required": [
          "valorLimite"
        ],
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "categoria": {
            "type": "string",
            "enum": [
              "COMBUSTIVEL",
              "MANUTENCAO",
              "SEGURO",
              "IPVA",
              "MULTA",
              "PEDAGIO",
              "LAVAGEM",
              "OUTRO"
            ]
          },
          "valorLimite": {
            "minimum": 0.01,
            "exclusiveMinimum": false,
            "type": "number"
          }
        }
      },
      "BudgetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "categoria": {
            "type": "string",
            "enum": [
              "COMBUSTIVEL",
              "MANUTENCAO",
              "SEGURO",
              "IPVA",
              "MULTA",
              "PEDAGIO",
              "LAVAGEM",
              "OUTRO"
            ]
          },
          "periodo": {
            "type": "string"
          },
          "valorLimite": {
            "type": "number"
          },
          "valorConsumido": {
            "type": "number"
          },
          "percentualConsumido": {
            "type": "number"
          }
        }
      },
...
    }
  }
}